Add command 'clean packages' to rmt-cli
RMT users have been asking for a way to clean stale packages in RMT,
which consumes extra disk space.

There are an open issue (https://github.com/SUSE/rmt/issues/662) and
an internal bug report requesting the feature.

Here, I implement the command 'rmt-cli clean packages', with the options
'--verbose', '--dry-run' and --non-interactive'. It's an initial
version, there are some improvements to be made based on
https://github.com/ikapelyukhin/repo-tools/blob/master/clean_outdated.rb:

- Checking inodes to correctly calculate file sizes;
- Delete only files which are two or more days old.